Address PR review feedback across scheduler components and CMake portability
Agent-Logs-Url: https://github.com/jeffreyqdd/vortex/sessions/8c264292-81eb-4e84-8871-bce4a928f9de

+++ src/scheduler-components/include/vortex_scheduler/ingress_types.hpp
@@ -11,7 +11,7 @@ VORTEX_SCHEDULER_NAMESPACE_BEGIN
 // Internal runtime handle for arena-owned payload bytes.
 struct BlobHandle {
   uint8_t pool_class = 0;
-  uint32_t segment_id = 0;
+  uint64_t segment_id = 0;
   uint32_t offset = 0;
   uint32_t size = 0;
 };
